+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 #Changelog
 All Notable changes to `Fsilva\HttpMessage` will be documented in this file
 
+## NEXT - YYYY-MM-DD
+
+### Added
+- Buffer stream
+
 ## 0.1.0 - 2015-02-28
 
 ### Added